third_party/blink/web_tests/virtual/web-components-v0-disabled/shadow-dom/range-caret-range-from-point-left-of-shadow-expected.txt

CONSOLE ERROR: line 29: Uncaught TypeError: parent.createShadowRoot is not a function
This is a testharness.js-based test.
FAIL range-caret-range-from-point-left-of-shadow Uncaught TypeError: parent.createShadowRoot is not a function
Harness: the test ran to completion.
